todo list app with suer authentication

+++ server/controllers/todoController.js
@@ -3,7 +3,10 @@ const Todo = require('../models/Todo');
 // Get all todos
 exports.getAllTodos = async (req, res) => {
   try {
-    const todos = await Todo.find();
+    const { userId } = req.query;
+    console.log(userId)
+    const todos = await Todo.find({ userId });
+    // console.log(todos)
     res.json(todos);
     // console.log(todos)
   } catch (error) {
@@ -15,8 +18,10 @@ exports.getAllTodos = async (req, res) => {
 exports.createTodo = async (req, res) => {
   try {
     const { task } = req.body;
+    const { userId } = req.query;
     const todo = new Todo({
       task,
+      userId: userId,
     });
     await todo.save();
     res.status(201).json(todo);
@@ -37,6 +42,7 @@ exports.updateTodo = async (req, res) => {
       { task, completed, completed_time },
       { new: true }
     );
+    // console.log(todo)
     res.json(todo);
   } catch (error) {
     res.status(500).json({ error: 'Server error' });
@@ -47,7 +53,7 @@ exports.updateTodo = async (req, res) => {
 exports.deleteTodo = async (req, res) => {
   try {
     const { id } = req.params;
-    await Todo.findByIdAndDelete(id);
+    await Todo.findOneAndDelete({ _id: id});
     res.json({ message: 'Todo deleted successfully' });
   } catch (error) {
     res.status(500).json({ error: 'Server error' });
@@ -57,10 +63,10 @@ exports.deleteTodo = async (req, res) => {
 // Delete all todos
 exports.deleteAllTodos = async (req, res) => {
     try {
-        await Todo.deleteMany();
+        const { userId } = req.query;
+        await Todo.deleteMany({ userId });
         res.json({ message: 'All todos deleted successfully', status: 200 });
     } catch (error) {
         res.status(500).json({ message: 'Server error', status: 500 });
     }
 };
-
